Refactor Layer: split layers into 2 classes
The "Layer" class and its subclasses are only responsible for
creating geometries out of the data, with the geometrical objects
as the final result obtained from this class, along with metadata.

The "Geometry" (so called because Renderable is already taken) takes
geometrical objects and renders them. This can be re-used for other
plottables like Guides.