path change animation
see http://bost.ocks.org/mike/path/ for an explanation of the problem

we try to use mbostock's method of interpolating the transform rather
than interpolating the entire change. this means animating the scale
change rather than the total change of position for the pahtes.

this is not a perfect change (deletation should ideally happen before
animation), but it is a quicker fix than expected. the change affects
lines, splines, and step charts