Add heuristic to avoid problems with unix time stamps.
When dates are provided as Unix time stamps without a format meta,
we pass it on to Moment.js to parse. However, Moment assumes they
are in some form like 'YYYYMMDD' and produces a time stamp thusly.

Since current timestamps are more than 8 digits, we should be able
to assume that long strings are timestamps.

Change 8 -> 9 in heuristic.